backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
marshmallow
google-genai
groq
python-dotenv
